+++ src/spines/__init__.py
@@ -2,11 +2,17 @@
 """
 Spines
 
-Backbones for parameterized models.
+Skeletons for parameterized models.
 """
-from .models import Model
+from ._version import get_versions
+__version__ = get_versions()['version']
+del get_versions
+
 from .parameters import Parameter, HyperParameter
 from .parameters import Bounded, HyperBounded
+from . import transforms
+from .model import Model
+from . import utils
 
 __all__ = [
     # Models
@@ -16,8 +22,7 @@ __all__ = [
     'HyperParameter',
     'Bounded',
     'HyperBounded',
+    # Submodules
+    'transforms',
+    'utils',
 ]
-
-from ._version import get_versions
-__version__ = get_versions()['version']
-del get_versions