+++ src/spines/parameters/__init__.py
@@ -1,18 +1,14 @@
 # -*- coding: utf-8 -*-
 """
-Parameters module for parameterized models.
+Parameters module for spines.
 """
 from .base import Parameter
 from .base import HyperParameter
-
 from .base import InvalidParameterException
 from .base import MissingParameterException
-
 from .core import Bounded
 from .core import HyperBounded
-
 from .store import ParameterStore
-
 
 __all__ = [
     # Parameters