+++ src/spines/parameters/store.py
@@ -6,37 +6,13 @@ Parameter storage module.
 #   Imports
 #
 from collections.abc import MutableMapping
-from functools import wraps
-from typing import Dict, Iterator, Type
+from typing import Dict
+from typing import Iterator
+from typing import Type
 
 from .base import Parameter
 from .base import MissingParameterException
-
-
-#
-#   Decorators
-#
-
-def state_changed(func):
-    """Decorator indicating a function which changes the state
-
-    Parameters
-    ----------
-    func : callable
-        The function to wrap.
-
-    Returns
-    -------
-    callable
-        The wrapped function.
-
-    """
-    @wraps(func)
-    def wrapped(self, *args, **kwargs):
-        ret = func(self, *args, **kwargs)
-        self._finalized = False
-        return ret
-    return wrapped
+from .decorators import state_changed
 
 
 #
@@ -51,10 +27,8 @@ class ParameterStore(MutableMapping):
     def __init__(self):
         self._params = dict()
         self._values = dict()
-
         self._finalized = True
-
-    # dunder methods
+        return
 
     def __repr__(self):
         ret = "<%s final=%s> {\n" % (self.__class__.__name__, self._finalized)
@@ -69,8 +43,6 @@ class ParameterStore(MutableMapping):
             ret += "  %s: %s\n" % (v, self._values.get(k))
         ret += "}"
         return ret
-
-    # Mapping functions
 
     @state_changed
     def __setitem__(self, k: str, v) -> None:
@@ -89,8 +61,6 @@ class ParameterStore(MutableMapping):
 
     def __iter__(self) -> Iterator[str]:
         return iter(self._values)
-
-    # Properties
 
     @property
     def parameters(self) -> Dict[str, Parameter]:
@@ -111,8 +81,6 @@ class ParameterStore(MutableMapping):
     def final(self) -> bool:
         """bool: Whethor or not this set of parameters is finalized."""
         return self._finalized
-
-    # Helper methods
 
     def copy(self, deep: bool = False) -> Type['ParameterStore']:
         """Returns a copy of this parameter store object.
@@ -143,15 +111,13 @@ class ParameterStore(MutableMapping):
         self._values.clear()
         self._params.clear()
 
-    # Option methods
-
     @state_changed
     def add(self, parameter: Type[Parameter]) -> None:
         """Add a :class:`Parameter` specification to this store
 
         Parameters
         ----------
-        option : Parameter
+        parameter : Parameter
             :class:`Parameter` specification to add to this parameter
             store.