+++ tasks/__init__.py
@@ -5,12 +5,15 @@ Invoke commands for common tasks.
 #
 #   Imports
 #
+import dotenv
 import invoke
 
+from . import develop
+from . import docs
 from . import install
 from . import release
 from . import uninstall
 
+dotenv.load_dotenv()
 
-namespace = invoke.Collection(install, release, uninstall)
-
+namespace = invoke.Collection(develop, docs, install, release, uninstall)